+++ src/convert/index.ts
@@ -1,2 +1,3 @@
 export * from "./optional-string.js";
 export * from "./string.js";
+export * from "./trim.js";