+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 19.7
 + Add optional `source` field to signals.
 + Add lifecycle support to `created` hooks.
++ Add `poll` utility to `rvx/test`.
 
 ## 19.6
 + Use exact prototypes for wrapping objects in `rvx/store`.