+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 19.7
 + **Deprecated:** `created` will be removed in the next major version. `created` is now an alias for `useMicrotask` and has lifecycle support.
++ **Deprecated:** `string` & `optionalString` have been moved to `rvx/convert` and will be removed from the core module in the next major version.
 + Add optional `source` field to signals.
 + Add `poll` test utility.
 + Add `useMicrotask` utility.